Add Google account email to application edit and show pages.
Currently, after a new member is accepted, they need to add a
Google-friendly email address that we can use for sharing the DU Google
Drive and Google Calendar. The membership coordinators often need to
remind people to add Google-friendly emails.

This PR moves adding a Google-friendly email into the application
process. The Google-friendly email is optional as suggested by Britta in
issue #487, in case someone wants to join without a Google address.