Move block of social media links to the bottom of the application form.
As suggested in https://github.com/doubleunion/arooo/issues/562.

I also removed the nested-table style from the stylesheet for this page.
It was only used for the table that contains the checkboxes at the end
of the form. I think the table with the checkboxes actually looks
slightly better without this style.