Remove configurable for orientation signup link.
We have a rolling application process now, so we won't do one batch of
orientations all at once that all new members should sign up for.
I think it would be hard to keep an orientation signup page up to date as
new members apply throughout the year, and it would probably be easier
to have the membership coordinator handle scheduling orientations for
new members as needed.